requirements.txt
fastapi
uvicorn
google-generativeai
google-cloud-aiplatform
pydantic